requirements.txt
comfyui
torch
importlib_metadata
huggingface_hub
scipy
opencv-python>=4.7.0.72
filelock
numpy
Pillow
einops
torchvision
pyyaml
scikit-image
python-dateutil
mediapipe
svglib
fvcore
yapf
omegaconf
ftfy
addict
yacs
trimesh[easy]
albumentations
scikit-learn
onnxruntime-gpu @ https://aiinfra.pkgs.visualstudio.com/2692857e-05ef-43b4-ba9c-ccf1c22c437c/_packaging/9387c3aa-d9ad-4513-968c-383f6f7f53b8/pypi/download/onnxruntime-gpu/1.17/onnxruntime_gpu-1.17.0-cp311-cp311-win_amd64.whl ;platform_system == 'Windows'